// DreamLuso.Presentation/DreamLuso.Presentation/src/app/models/CreateRealEstateAgent.ts
export interface CreateRealEstateAgent {
  userId: string;
  officeEmail: string;
  totalSales: number;
  totalListings: number;
  certifications: string[];
  languagesSpoken: Languages[];
}

export enum Languages {
  English = 'English',
  Spanish = 'Spanish',
  French = 'French',
  German = 'German',
  Italian = 'Italian',
  Portuguese = 'Portuguese'
}